Keep in-flight dedupe out of the core bundle
Move getStatisticIds to its own module so core.ts no longer pulls
shareInFlightRequest into the legacy core entrypoint, and inline
deepFreeze to avoid the deep-freeze dependency.
