+++ src/data/recorder.ts
@@ -3,7 +3,6 @@ import { computeStateName } from "../common/entity/compute_state_name";
 import type { HaDurationData } from "../components/ha-duration-input";
 import type { HomeAssistant } from "../types";
 import { firstWeekday } from "../common/datetime/first_weekday";
-import { shareInFlightRequest } from "../common/util/share-in-flight-request";
 
 export interface RecorderInfo {
   backlog: number | null;
@@ -160,24 +159,6 @@ export const getRecorderEntityOptions = (
     type: "recorder/entity_options/get",
     entity_id,
   });
-
-type StatisticIdsType = "mean" | "sum";
-
-type StatisticIdsApi = Pick<HomeAssistant, "callWS">;
-
-export const getStatisticIds = (
-  hass: StatisticIdsApi,
-  statistic_type?: StatisticIdsType
-) =>
-  shareInFlightRequest(
-    hass.callWS,
-    `recorder/list_statistic_ids:${statistic_type ?? "all"}`,
-    () =>
-      hass.callWS<StatisticsMetaData[]>({
-        type: "recorder/list_statistic_ids",
-        statistic_type,
-      })
-  );
 
 export const getStatisticMetadata = (
   hass: HomeAssistant,
